Move fetching canteens for diag creation and prefill to actions view
